requirements.txt
datetime
flask
flask_cors
numpy
pillow
flask_pymongo
python-dotenv
requests